+++ r-network.tf
@@ -18,8 +18,8 @@ module "subnet_gateway" {
   subnet_cidr_list     = var.subnet_gateway_cidr
 }
 
-resource "azurerm_virtual_network_gateway" "er_gateway" {
-  for_each = toset(var.express_route_gateway_enabled ? ["vgw"] : [])
+resource "azurerm_virtual_network_gateway" "ergw" {
+  for_each = toset(var.express_route_gateway_enabled ? ["ergw"] : [])
 
   name = local.vgw_name
 
@@ -31,10 +31,10 @@ resource "azurerm_virtual_network_gateway" "er_gateway" {
   sku           = var.express_route_gateway_sku
 
   dynamic "ip_configuration" {
-    for_each = var.express_route_gateway_sku != "Standard" && var.active_active_enabled ? ["1", "2"] : ["1"]
+    for_each = local.public_ip_number
 
     content {
-      name                 = "${local.express_route_gateway_ipconfig_name}-0${ip_configuration.value}"
+      name                 = format("%s%s", local.express_route_gateway_ipconfig_name, length(local.public_ip_number) > 1 ? "-0${ip_configuration.value}" : "")
       public_ip_address_id = azurerm_public_ip.public_ip[ip_configuration.value].id
       subnet_id            = var.subnet_gateway_cidr != null ? module.subnet_gateway["subnet_gateway"].subnet_id : var.subnet_gateway_id
     }
@@ -45,7 +45,7 @@ resource "azurerm_virtual_network_gateway" "er_gateway" {
 resource "azurerm_public_ip" "public_ip" {
   for_each = toset(var.express_route_gateway_enabled ? local.public_ip_number : [])
 
-  name                = "${local.pub_ip_name}-0${each.value}"
+  name                = format("%s%s", local.pub_ip_name, length(local.public_ip_number) > 1 ? "-0${each.value}" : "")
   location            = var.location
   resource_group_name = var.resource_group_name
   allocation_method   = var.public_ip_allocation_method
@@ -65,7 +65,7 @@ resource "azurerm_virtual_network_gateway_connection" "er_gateway_connection" {
 
   type                           = "ExpressRoute"
   express_route_circuit_id       = azurerm_express_route_circuit.erc.id
-  virtual_network_gateway_id     = azurerm_virtual_network_gateway.er_gateway["vgw"].id
+  virtual_network_gateway_id     = azurerm_virtual_network_gateway.ergw["ergw"].id
   local_azure_ip_address_enabled = false
   routing_weight                 = var.express_route_gateway_connection_route_weight