+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # Unreleased
 
 Added
-	* AZ-780: First release of ExpressRoute module
+  * AZ-780: First release of ExpressRoute module